Convert ShardingRuleConfiguration to DistSQL via SPI
